+++ packages/react/Chip/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The Chip component is primarily used to define cumulative criteria or attributes for searching or filtering information in a list of entries or objects. Each of these criteria can be added or removed using this component.
 
+## 2022-09-24
+
+### 💡 Others
+
+- Adjusted `fontSize` size of internal `Text` component from `small` to `caption`. ([#29](https://github.com/TiendaNube/nimbus-design-system/pull/29) by [@juniorconquista](https://github.com/juniorconquista))
+
 ## 2022-09-01
 
 ### 📚 3rd party library updates